backend/requirements.txt
openai
azure-core
python-dotenv
fastapi
uvicorn